requirements.txt
# Core dependencies
openai>=1.0.0
neo4j>=5.0.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# NLP analysis
jieba>=0.42.1

# Utilities
requests>=2.31.0
tqdm>=4.65.0

# Visualization (optional)
matplotlib>=3.7.0
pyvis>=0.3.2

